requirements.txt
python-telegram-bot
python-dotenv
fastapi
uvicorn
httpx